Now convert automatic all the java files of a given folder. Refactoring package.
