+++ pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>de.aschoerk.java-converter</groupId>
-  <artifactId>java-to-rust</artifactId>
+  <artifactId>java-2-rust</artifactId>
   <version>1.0-SNAPSHOT</version>
   <name>java to rust</name>
   <url>http://maven.apache.org</url>
