+++ src/main/java/de/aschoerk/java2rust/JavaConverterData.java
@@ -1,0 +1,7 @@
+package de.aschoerk.java2rust;
+
+/**
+ * Created by aschoerk on 30.04.16.
+ */
+public class JavaConverterData {
+}
