+++ src/main/java/de/aschoerk/java2rust/TrackedNodes.java
@@ -1,0 +1,8 @@
+package de.aschoerk.java2rust;
+
+/**
+ * @author aschoerk
+ */
+public class TrackedNodes {
+
+}
