--- src/main/java/de/aschoerk/javaconv/JavaConverterData.java
@@ -1,7 +1,0 @@
-package de.aschoerk.javaconv;
-
-/**
- * Created by aschoerk on 30.04.16.
- */
-public class JavaConverterData {
-}
