--- src/main/java/de/aschoerk/javaconv/TrackedNodes.java
@@ -1,8 +1,0 @@
-package de.aschoerk.javaconv;
-
-/**
- * @author aschoerk
- */
-public class TrackedNodes {
-
-}
