--- src/main/java/de/aschoerk/javaconv/TypeDescription.java
@@ -1,22 +1,0 @@
-package de.aschoerk.javaconv;
-
-/**
- * @author aschoerk
- */
-public class TypeDescription {
-    Class clazz;
-    int arrayCount;
-
-    public TypeDescription(final int arrayCount, final Class clazz) {
-        this.arrayCount = arrayCount;
-        this.clazz = clazz;
-    }
-
-    public int getArrayCount() {
-        return arrayCount;
-    }
-
-    public Class getClazz() {
-        return clazz;
-    }
-}
